refactor: 로또 서비스 로직 분리
